+++ src/sql/transactionOutputs.sql
@@ -1,7 +1,7 @@
 SELECT 
 tx_hash AS "txId",
-txout_pos::bigint AS "n",
+txout_pos AS "n",
 txout_scriptpubkey AS "scriptPubKey",
-txout_value::bigint AS "value"
+txout_value AS "value"
 FROM txout_view
 WHERE tx_hash IN ({{txIds | join(',')}})
